[DTFx/AzureServiceFabric] Interface to fetch running orchestrations and runtimestate
